Software_Projects/plaid-assistant/ai-fin-tracker/requirements.txt
python-dotenv
plaid-python
